// src/app/models/car-model.ts
export class CarModel {
  // tslint:disable-next-line: variable-name
  _id: string;
  ref: string;
  name: string;
  slug: string;
  // tslint:disable-next-line: variable-name
  desc_excerpt: string;
  description: string;
  // tslint:disable-next-line: variable-name
  date_online: string;
  // tslint:disable-next-line: variable-name
  date_offline: string;
  currency: string;
  // tslint:disable-next-line: variable-name
  contact_phone: string;
  // tslint:disable-next-line: variable-name
  contact_email: string;
  // tslint:disable-next-line: variable-name
  body_type: string;
  engine: string;
  mileage: string;
  // tslint:disable-next-line: variable-name
  fuel_type: string;
  transmission: string;
  // tslint:disable-next-line: variable-name
  door_count: string;
  // tslint:disable-next-line: variable-name
  image_car: string;
  price: number;
}